working initialization of nutrients and targets. weights and progress set to 0. DB int64 successfully changed to Double

+++ Nutrient+/NutrientPlus/NutrientStruct.swift
@@ -10,9 +10,9 @@ import Foundation
 
 class NutrientStruct {
     var nutrName: String
-    var nutrWeight: Int64
-    var nutrTarget: Int64
-    var nutrProgress: Int64
+    var nutrWeight: Double
+    var nutrTarget: Double
+    var nutrProgress: Double
 
     init(nutrName: String) {
         self.nutrName = nutrName
@@ -21,7 +21,7 @@ class NutrientStruct {
         nutrProgress = 0
     }
 
-    init(nutrName: String, nutrWeight: Int64, nutrTarget: Int64, nutrProgress: Int64) {
+    init(nutrName: String, nutrWeight: Double, nutrTarget: Double, nutrProgress: Double) {
         self.nutrName = nutrName
         self.nutrWeight = nutrWeight
         self.nutrTarget = nutrTarget
